+++ lib/crypto/errors.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2021 Carl Schwan <carlschwan@kde.org>
 // SPDX-License-Identifier: LGPL-2.1-or-later
 #ifdef Quotient_E2EE_ENABLED
-#include "olm/errors.h"
+#include "crypto/errors.h"
 
 Quotient::OlmError Quotient::fromString(const std::string &error_raw) {
     if (error_raw.compare("BAD_ACCOUNT_KEY")) {
