+++ lib/crypto/qolmaccount.h
@@ -4,10 +4,10 @@
 #pragma once
 #ifdef Quotient_E2EE_ENABLED
 
-#include "olm/e2ee.h"
-#include "olm/errors.h"
-#include "olm/olm.h"
-#include "olm/session.h"
+#include "crypto/e2ee.h"
+#include "crypto/errors.h"
+#include "crypto/session.h"
+#include <olm/olm.h>
 #include <QObject>
 
 struct OlmAccount;
