+++ lib/crypto/qolmsession.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-#include "olm/qolmsession.h"
+#include "crypto/qolmsession.h"
 
 using namespace Quotient;
